Add files via upload
This script will take 2 CSV files and compare a column between them. It is best if the column to be compared in the secondary one contains unique values so that it more accurately appends the corresponding row data to the primary.

For example: A vuln scan with a column showing IPs as a primary and an asset inventory showing IPs as a secondary. The asset inventory data for each host can be used to enrich the vuln scan data.