Update BulkWhois.py
Now leveraging the ipaddress module to more accurately identify public IP addresses.

Redid the way the data is stored for use by the script.

Added a "quick search" feature that if used (yes by default), for the IPs you're wanting to search it'll iterate through the whois data already pulled to match IPs with CIDRs. This may help in times when individual queries are taking longer than expected.