Update cidrFromRange.py
Added functionality for simpler more collapsed subnets to be returned. Sometimes this is beneficial if your dump has a range of, for example 10.1.1.1-10.1.1.254; the expanded option will truly give you each subnet, even if there's a /32 in there, where the not expanded version will just present you with a single /24. Hey, just add the -e option and you can get all of the crazy subnets!